Create AssetRecordDecorator to assist with routing
As multiple models need to use the asset record partial, it creates a
routing problem as no single rails routing helper can be used.

Instead an AssetRecordDecorator has been created that forms the route
based off the model name and id. This decorator assumes that all models
that have an asset_record can be routed to
:mode-name-plural/:id/asset_record

The decorator is extended in the ApplicationDecorator by checking if the
model responds_to? :asset_record. This means the decorator does not need
to be explicitly included in each models decorator.

Instead including the HasAssetRecord concern in the model will
automagically trigger the decorator to be activated.

This means the routes no longer need to be passed into the partial
implicitly.