Redirect to the asset page after update asset record
The update method has been added to the AssetRecordsController. The
final action the controller needs to do is redirect to the asset page.

Currently the data is not being updated, this will be done in a latter
commit.

The AssetRecordDecorator needed to be adjusted to use a hyphen instead
of a underscore in the path. This will need to be the convention when
using multiword models with an AssetRecord.