Merge asset_records as a hash before converting to array
Previously the inject method required mutating the array which caused
bugs with the rails caching.

Instead the asset_record is stored as a hash based on the definition id.
This hash can then be used to merge the parent record cleanly without
mutating the object.

The `asset_record` method then converts the hash into an array to be
used by the views.