Delete the asset record when it is updated to be blank
If the updated value is an empty string or nil, then this triggers the
record to be deleted. This is so the record can revert to its higher
level.

It may not be possible for a nil value to be set as a field, however it
is included for robustness.