Link AssetRecordsController to the update code
The controller now calls the code to update the asset record. A few
minor changes where required as HasAssetRecord expects a hash.

Also additional code has been added to insure the keys are symbolizes.
Otherwise the keys could be a symbol, string, or integer.

NOTE: symbolize_keys can not be used to symbolize integers. This is
because integers do not respond to to_sym. Instead they must be
converted to a string first.