Fix bug when deleting higher level asset fields
If a higher level asset sets a field but then the field is deleted, it
shouldn't do anything. Previously it would save a new blank entry at the
current level. It doesn't affect the parent entry.

This Previously wasn't a problem, but it is intended to only save the
parent field as a placeholder. If this is the case, then forms with
blank fields will be submitted regularly but this does not mean replace
the field with a blank.