Allow overriding higher level assets w/ same value
Now that the form uses place holders to display higher level assets
values it will no longer submit them to the app. This means the update
no longer needs to filter out values that match the existing value (i.e.
the higher level).

This also allows a value to be specifically set at the lower level even
if it matches it's parent.

Due to this change, the spec needed to be updated so it doesn't send the
old hash, but instead sends the updated values.

Not the do nothing test was removed as the behaviour is now undefined.
If a blank value is received it could me do nothing (for higher levels)
or delete if it is set at the current level. Note: values for the
current level are still pre-filled and will be sent with the update
request.