Display a select box to change the ComponentMake
As part of the AssetRecord partial, a new field has been added to
change the ComponentMake. It can only be changed when editing a
ComponentGroup, otherwise the value is displayed.

Inorder to display something meaningful about the make, a name method
has been added to the model that takes the manufacture and appends the
model to it.