Allow ComponentGroup to update the ComponentMake
The AssetRecordController has been edited to allow the ComponentMake to
be updated. The can only occur for ComponentGroup as Components do not
directly link to the make.

There is a check to prevent the ComponentMake to be changed to a
different ComponentType. In practice this shouldn't be possible as the
form only shows valid options. However it is good to have the check
server side.