Extract creating asset form inputs into decorator
The construction of the form input is a complex process and is better
suited in a Decorator than the view it self.

Also displaying the table depending on the action has been simplified.
First an enumerator is constructed based on the action and then the
enum is used to create the table.