Require an explicit empty string to delete an record
During testing I found that if the AssetRecordTable form has a bug and
doesn't submit values with the correct definition ids, then it deletes
the old data.

The reason is the incorrect parameters to the controller are ignored,
but an empty hash of "correct" parameters is still made. This empty hash
always returns nil values which trigger the asset_records to be deleted.

Instead a nil should be ignored as it is more indicative of missing
data. Instead an empty string needs to be explicitly returned from the
form in order to trigger a delete.