Add test to check data_type issues the Deprecation msg
As the production database does not contain the data_types fields, it
the data_type method needs to be able to handle nills.

In this case, it should default to "short_text". A deprecation warning
has been added to notify that the data_type has been implied.

Once the production db has been updated, then the default/ warning can
be removed.