Seed the database with 'long_text' data_type
In development, we need some of the data to be 'long_text' data_type.
This commit modifies the seed data itself. This should really be in the
development seed data however the production db shouldn't be reseeded
with this data.

This should be fine for know, however we should pull the data from the
production db for the global tables (e.g. AssetRecordFieldDefinitions)