Create AssetRecordFieldDecorator to contain shared code
The options block is highly dependent on the ARField and current asset
and thus should not be stored in the ARFDefDecorator which will be
overridden with different types.

Instead the options and value will be determined in a ARFieldDecorator
and then passed onto the Definition.

The current_asset still needs to be passed in as this can not be
inferred from AssetRecordField.