Add subject to Cases, use Issue default if unset
This will hold the main part of the subject used when creating an RT
ticket, which will now be user-editable (rather than always set to the
Issue name) and so needs to be saved with the other data on a Case.

The entire subject used when creating an RT ticket will not be
user-editable or saved in this field, as we always want to include
certain details, in particular a prefix with the Cluster name and a
suffix containing the token for disambiguating Cases.