Migrate existing Cases to have new fields set
As well as slightly tweaking how this token is generated (while
retaining the same effect), to reflect the fact that this should always
be generated on Case creation.

In particular setting a token for every Case is needed as otherwise,
every time a Case is then read (until it is saved again, which doesn't
normally happen AFAICR) a new token is created, meaning using the
`mailto` link multiple times for a Case would not lead to these emails
being threaded.

Setting a subject for every Case is less essential as the default
subject would currently always be the same, but is still important
firstly as what this default is might change, and secondly as the
subject is currently only set to the default just before validation
(this behaviour could probably change, but we might as well do the thing
most likely to be robust).