Display banner indicating when site in development
Now that we are importing production database backups to use as our
development data, it will probably be useful to have a prominent way to
distinguish whether we are viewing a development or production version
of the site. This will probably be most important to try and avoid
anyone accidentally doing something which will affect users on the
production site thinking it is their local version.