Validate that Definition data_types' are valid
ARFDefinition now validates that the data_type it is storing has been
previously whitelisted in ARFields::VALID_DATA_TYPES

Also fixed broken tests. The warning needs to be called at least once
not exactly once. The validation is throwing an error on validation.
Curious that it occurs on `create` not `create!`, however switching the
spec to `build` fixes it. This shouldn't be an issue as types are not
going to be added often.