Correctly format title text following rebase
Following a rebase the new format for the Cases table header was no
longer being correctly formatted, due to it now being passed through the
`card_header_nav` partial, causing `&mdash;` to be treated as a literal
string rather than a HTML entity. Wrapping the whole title in `raw`
fixes this.