Filter the updated records for failures
The updated asset records array contains a combination of valid and
invalid records and nils. The nils represent when nothing happened or
when a delete successfully occurred. If the delete fails than an error
500 will still occur. However currently the delete shouldn't fail.

The array is therefore filtered to only contain invalid records which
are added to an error_objects array. This replaces the error_flag.

Eventually the error_objects will be used to issue the flash