Add migration to add back missing Issue Categories
We recently changed Issues to be related to Services, and make Issue
Categories not required (in order to make the Case form read more
intuitively and not include unnecessary Category fields containing only
one or two Issues). At some point in this process we lost the
relationship between Issues and Categories that we did still want to
keep, to split up the Issues for the 'HPC Environment' Service as there
are quite a lot of these. This migration adds these back appropriately.