Hack out admin config asset record editing
This is now a worse version (in terms of both code and the UI) of the
asset record editing which has now been added for Components and
ComponentGroups within the main app, which is also only accessible to
admins. Therefore we might as well remove this functionality from the
dedicated admin interface.