Extract couple of CaseDecorator methods
We have a couple of columns in the Cases table containing logic we would
like to share with the new pages for individual Cases (to be added
soon). The logical place for this therefore is the CaseDecorator.

In particular extracting the `credit_charge_info` method allows us to
tidy up where this was previously used in the Cases table, and add some
fine-grained test coverage for this to the CaseDecorator specs.