Remove very old broken test
This test is quite old and broke a while ago:
3b26bacd712bd98567ca5e7f5eb65fe087b132a1

However this file wasn't named correctly and thus was being skipped. The
test only started running when the final name was changed.

Due to changes in the controller, this test no-longer fails. However it
only concerns the admin side of the website. Thus we do not care if a
rogue POST would trigger an incorrect action. It would require an admin
login who already has complete privileges.