Add initial page for individual Cases
Currently this just shows a card with a table showing all important
details for a Case. This will allow us to make the main Case table less
large and busy, as well as give us a page to link to in future for each
Case and to put additional information related to a Case (such as
related maintenance windows).