Work around Rails implicitly running tasks for test db
Rails implicitly always also runs `db:*` tasks in the `test` environment
when they are run for the `development` environment, even if you try to
explicitly specify that they should only run in the `development`
environment. Re-loading the schema for the test database after it does
this at least leaves things in a usable state again. Relevant issue:
https://github.com/rails/rails/issues/27299.