Show subject with link to Case page in Cases table
When showing real production data (i.e. as it is shown to our real
users) the Cases table was becoming very long and busy. This is due both
to the amount of information included in the details for a real Case and
the frequency that real Cases are created.

Now that we have a separate Case page for each Case we can link to this
and just show each Case's subject in the table, which makes the table
much shorter and clearer. We might as well also no longer show the Issue
name in the Case table, since this is now always either the same or less
specific than the Case subject.