Create data migration for ComponentMakes and Expansions
As the relationship between ComponentGroup and ComponentType has been
broken, a data migration is required to add the relationship back.

This relationship occurs through ComponentMake. This migration contains
the new ComponentMakes and the associated DefaultExpansions. Thus the
ExpansionTypes also needed to be added.

This commit builds up to the first ComponentMake the '1U Server'.
Additional makes will be added in future commits.

NOTE: The `raise 'Exited correctly'` prevents the migration from
completing. This is intentional as the migration is not yet complete and
it prevents the db from needing to be rolled back.