Add the 'Network switch' and 'Disk array' makes
This uses the same method as the other makes. The code has not been
generalised as it is clearer what it happening when repeated. This
shouldn't be a huge issue as the migration code isn't likely to be
reused again.

Also changed 'Component Expansion' make to 'Component expansion' for
consistency.