Add validation that engineers are admins
The "engineer" that is creating the log must be a `User` with `admin`
privileges.

This is enforced using an `engineer_is_admin` validation that checks if
the engineer is an `admin?`. It does need to handle when the engineer is
`nil` otherwise it will error. The other validation will still detect it
is missing.