Ensure all cases belong to the correct cluster
There is two paths between Cluster and ClusterLog. The first is the
direct 1-many relationship. The second ("in theory") is a many-many
relationship via Case.

However in practice all the cases associated with ClusterLog must belong
to the same cluster. This preserves the initial 1-many relationship.

This is done by adding a validator in `ClusterLog` that checks the cases
belong to the correct cluster.