Update script to skip invalid columns
Some migration lines may be included in the output of this script which
we know will always be invalid and unnecessary, therefore hard-code
these to be skipped on future runs.

Also add a note to update the lines to skip if necessary, and abort the
script early when no new constraints needed.