Format form for creating new logs
The details and submit fields have been added to the form. The engineer
who is submitting the log can be inferred server side as the
current_user.

The form has been correctly formatted using bootstrap and labels.

There was an error where the form was for a new `Case` not `ClusterLog`.
This now has been fixed