Form the parameter hash for creating a log
The submitted parameters need to be processed before they can be used
for the update. The engineer is not submitted with the form as it is
always the current user.

The submitted case_ids also need to be processed as they contain an
empty string that is submitted as a placeholder.