Implement initial maintenance window states
This commit adds the state_machines-activerecord Gem for handling state
machines in Ruby and Active Record, and adjusts the MaintenanceWindow
model to use this. This is done by adding a `state` column to it to hold
this state, as well as defining the current states we have for
MaintenaceWindows (which were previously implicit) and how these should
be transitioned between.