Use state query methods to choose window icon
Rather than relying on previous methods used here which determine the
same thing but now in a roundabout way. Also note `in_progress?` method
introduced as although this is currently identical to `confirmed?`,
later it won't be and we will still want to handle this branch of the
conditional for the 'in progress' state.