Add MaintenanceWindow fields for new states
MaintenanceWindows are about to have several new states added, to
account for them now being scheduled for a particular time slot rather
than started as soon as confirmed by a user and ended manually by an
admin, as well as to better model the different actions we want to
support for admins and site contacts. To support this, we want to have
have the following new fields:

- `requested_start` and `requested_end`, for the requested scheduled
time period for the MaintenanceWindow;

- `${state}_by` fields, for every user-initiated state transition;

- `${state}_at` fields, for every state transition.

This commit adds each of these fields which doesn't exist yet; actually
having these be used anywhere is still to come in future commits.