Remove additional base Bootstrap import
We already now import all Bootstrap CSS, but with the custom theme that
we want, in app/assets/stylesheets/flight-branding/flight-chrome.css.

Therefore we can remove this import of the base Bootstrap CSS, which
reduces the amount of unnecessary CSS we need to load, avoids the flash
of different styling when one of these has loaded and the other hasn't,
and avoids any other possible conflicts with having multiple CSS files
both effecting the same classes.