Correct import script for new backup name format
We are now taking multiple database snapshots a day, and so are saving
these to S3 in a longer format including the time. Unfortunately
previously our database import script just read the date of the most
recent database backup, which just happened to be the same as the file
name. This commit corrects this to actually use the file name.