Create a context for testing the cluster log
Now that a log can be for a cluster and a component, the feature spec
has been updated to include a context. This will make it easier to add
the component specs in the future.

The spec has also been updated to create some test components. The
cluster log page should display a select box to display these
components.

A random :component and :case has also been added. These random models
belong to a different cluster and thus should be ignored. They represent
all other DB entries and should not affect the tests.