Clean up the log_spec setup
1. Rename :details_input_id

The previous local variable had refer to the case. However this input
has nothing to do with the case ids and thus has been renamed.

2. Remove :case_details

This was only being used to create the :cases and thus can be merged
into one method

3. Remove :case_select_details

This local variable is shadowing a case method. Instead it has been
replaced with its own helper (`select_details_for`). It is intended this
helper will be used more in the future.

4. Create cases in a before block

:cases is an ambiguous variable name particular because it doesn't
include all the cases (it is not equal to `Cases.all`). Instead the
cases are defined in a :before block. The one location :cases was being
used has been replaced with `cluster.cases`

5. Create components in a before block

Similarly to :cases, the :components can be created in a before block

6. Extract method for filling in the log details

This needs to be done in multiple location but is essentially the same
each time.