Incorporate started state into MW transitions
It is no longer possible for a MaintenanceWindow to be transitioned
directly from 'confirmed' to 'ended', it must go from 'confirmed' to
'started' and then 'started' to 'ended'. Both 'started' and 'ended'
should be automatic transitions which occur when the requested_start and
requested_end times are respectively reached (this will be handled in a
future commit).