Add component logs to the current component
The log will be added against the current component (as defined by the
path). The id from the path has been used as this allows the controller
to determine which component to use.

In order for this to work, the Log model needed to be updated to ensure
the cluster has been assigned. This is because the `cluster` is required
but can not be inferred automatically when a component log is being
made.