Set up state_machines-audit_trail Gem for MWs
The number of fields for tracking the time a MaintenanceWindow
transitioned to a particular state, along with the user who instigated
this transition (only applicable for some states), as well as the
validations for when these fields should/should not be set and tests for
all of this, is causing an explosion of complexity and making it
difficult to be sure whether we are handling/consistently handling all
possibilities.

Instead, we will start using the state_machines-audit_trail Gem to keep
track of transitions between MaintenanceWindow states along with any
needed metadata, and only have core MaintenanceWindow properties stored
within the MaintenanceWindow itself. This should make the handling of
our current states and transitions easier, make it simpler to add future
states and transitions in a consistent way, and also means we can easily
handle things like returning to a previous state and tracking this
(which is not currently needed but may later be).

This commit sets up the Gem and needed model and table for this, without
actually using this anywhere else (to be done in a future commit(s)).