Add a test for the correct number of links
In the navigation bar there is a bunch of site specific navigation
links. These links are dynamic and depend which section of flight centre
the user is visiting.

These links are always within the first 'ul' item. (There are two `ul`
items, the second contains the logout button). The items themselves are
`li`. The admins have an additional `all sites` button.