Determine the cluster links based on context
Testing the nav bar is tricky because the cluster links are the same for
both admins and users. The difficult bit is the indices are different
because admins have an additional link in the second (1st index)
location.

As a way of generalizing it, a new variable :cluster_nav_links has been
defined based on the context. For admins, it removes the first two links
(the root and site links). For users, it removes only the first links
(the root link).

By doing so, cluster_nav_links should be the same regardless of context,
making it easier to test.