Test if the second admin link is two the sites
Currently these spec is over complicated, as it needs to test a few
different things. This should probably be simplified by moving it into a
separate context.

However it is testing if the user is an admin, then the second link
should be to the sites page. User however can not navigate to different
sites. This means their root page is their "site" page and thus do not
have a separate link to it.