Condense the tests to save on setup time
As the shared examples are run multiple times in different contexts, the
time it takes run grows rapidly.

To make the tests faster, they have been condensed to reduce there
numbers. This reduces the number of setups required but still tests the
same functionality.

This has made the tests about 30% faster.