Remove nav id and use a class selector
The id was added solely so it can be selected in the spec. However
capybara supports selecting objects based on the tag and class name.

NOTE: Capybara uses css style selectors, so the `nav.product-bar` refers
to all `nav` objects with the `product-bar` class. The `first` selector
method unsurprisingly selects the first object that means this criteria.