Remove assignment of nav variables
The navigation bar is no longer going to be formed by setting a bunch of
variables. Instead a partial will be used which will read a scope
variable.

This will free up the instance_variables which maybe set by controllers
which actually use them.

The `case` statement will remain for the time being as it makes sense to
set some of the variables if the site directly relates to them.

This commit will break the tests. These will be fixed in future commits.