Use a default set of navigation variables
Instead of manually passing in nil navigation variables, they have now
been set in a default. In most cases they should be nil so this prevents
code duplication and ensures they aren't accidentally set.

The scope nav variable should be set most of the time and thus should
equal the subject of the spec.