Assign the scope in the application controller
The scope is set as the "model" that is currently being visited by the
page. Please note, not all models support the scope and thus it more
refers the different parts of the cluster.

Also note, this is different again to `cluster_part` as this does not
include component_groups for legacy reasons.

We need to get better at naming, sorry