Only allow variables to be set to the subject
At this stage there is no reason why other models (than the subject)
should be set. If a model can be accessed from the subject, then that is
how it should be done.

By preventing other models from being set in the background, there is
less chance they will cause a bug into the future.