Set the site for contact type user root page
In order for the root page for a contact to be displayed correctly, the
site must be in scope. This is a break from how the rest of the scopes
have been set.

It is because contact's root page is the same as an admin site page.
This means it needs to set the site as the scope. Where the admin site
page has nothing in scope.

This edge case will be handled in the rendering of the navbar as the
first two links for admins (and first link only for contacts) are kinda
special.