Migrate existing MaintenanceWindows to new format
This data migration converts all existing MaintenanceWindows to look
roughly like they would have if they had been performed in the new
system for requesting maintenance, with all fields which would now need
to be set and all needed MaintenanceWindowStateTransition records
created. Where data did not previously exist (e.g. the `created_at`
fields for the transition records) a best guess has been made for
plausible values for this.

This should be sufficient so that we can handle the past maintenance in
a similar way to new maintenance, without needing to special case it.