Rename method to better indicate what it represents
It was pointed out in
https://github.com/alces-software/alces-flight-center/pull/78#discussion_r170564102
that this is slightly confusing alongside the `confirmed?` method, as
surely a maintenance window can only be `confirmed?` or `unconfirmed?`.

However `confirmed?` represents the specific `confirmed` state, and
`unconfirmed?` represented the states prior to this where the window is
not yet confirmed; additionally there are also the states other than
these where the window may or may not have been confirmed by a user, but
has progressed beyond a point where it can be confirmed (as it has been
cancelled/rejected/the maintenance should have already started etc).