Add spec to ensure the last nav item is active
Previously some complicated logic was used for each link to check if it
is active. This has since been removed as the variables used to set them
have been removed.

This functionality is going to be added back in a future commit. But
before this happens, the spec has been updated to ensure the last item
is active.

This should always be the case and should hopefully simplify the logic.