Render all nav links from within a proc
The problem with having only having the last link active is that
information isn't know in advance.

To get around this problem, then links shall be rendered from within a
proc. This allows an array of procs to be created and then it is a
trivial case of rendering the last proc as the active item.