Set the last link as the active link
This commit adds the code required to set the last link as active. The
proc has been updated so it can take any additional parameters. This
should improve readability and will allow additional parameters to be
added in the future (if required).

The `active` parameter is optional and defaults to false if it hasn't
been set. This means the non-last links do not need to provide if they
are active or not.