Correctly set the current_site for logged in users
A lot of code is dependent on knowing which site is currently being
used. Unfortunately this is tricky to determine as it depends on which
page is being used AND who is viewing it.

In certain cases the `site` variable has been set. These are the easiest
case as this variable should be returned.

Otherwise the current site is the one that corresponds to the scope.

For this to work correctly, every user site requires a scope (e.g. case
pages). This is why a full back to the `scope = current_user.site` has
been added. This only works for non-admin users as admins do not have a
site.

The edge case where the scope has not been set for admins is not
possible. This is because admins have the `/site/:id` paths which set
the scope, if required.

The condition where the scope for `/site/:id` paths are only set for
admins has been removed. The routing ensures admins are the only ones
that can view this site so it is superfluous.