Allow links to be render for models
The text and path for a particular model are typically quite straight
forward to determine.

So instead of manually inputting them in all cases, the model can be
passed instead. This allows them to be determined implicitly.

There are still cases where the path and text needs to be overridden and
thus that functionality has been preserved.