Render all the nav links in a single loop
Instead of slicing the proc array, it can be done within a single loop.
This can be done by setting the active variable in all cases, depending
if it is the last proc or not.

Also setting the default `active` variable is being set at the top of
the partial. This is still required as there are other links that use
the partial which are not active

https://github.com/alces-software/alces-flight-center/pull/86#discussion_r170904738