Create cluster tabs partial
This is the first step for creating tabs for the different sections of
the cluster page. Each section will be on its own page which can be
navigated with the tabs.

The tabs in turn use the nav_link partial from the header nav bar.
Because these links do not have icons, the `nav_icon` variable needs to
be optional.