Show cluster cases on a separate tab
A new route to the index action has been added for cluster cases. This
page is linked to from the main cluster page. NOTE: a back button has
not been added as it will eventually be replaced by tabs to all
sections.

The cases view index template already exists. Previously it only worked
for sites, however now it uses the new `scope` variable.

Also, if it is showing the cluster cases, it displays the cluster case
bar. This can probably be generalized in the future.

Currently the case form buttons are broken as they didn't work for the
cluster object.