Allow tab links to be activated by id
The current tab link (currently only cases) can be activated by and id
associated with the link. `activate` is used to pass the symbol so it
does not clash with the underlying `active` boolean used by the link.

As there is only one link that is implicitly non-active (the admin
interface link), this functionality has been removed. This links to a
comment give on:

https://github.com/alces-software/alces-flight-center/pull/86#discussion_r170905213