Add the case form table back
As the case_form_table uses a decorated methods AND is called from a lot
of places, it is bug prone to assume it has already been called.

Instead it decorates the model for itself.

In the future, the case_from_table should be removed and replaced with
the :index page for cases. The cases should be then set by the
controller and not the model.

This will be facilitated by the use of tabs and the scope.