Fix maintenance window spec to be less flaky
Previously the spec relied on the fact that the maintenance table was
the second table on the page.

This prevents the layout of the page from being altered, even though the
underling feature is still functional.

Instead the maintenance table has been added to the maintenance class.
This allows capybara to find the table directly without making an
assumption about the page layout.