Create helper method for rendering the tab bar
The tab bar needs to be rendered on multiple cluster pages. Currently,
there is only a cluster tab bar, but this will likely change in the
future.

This means the same template will need to render a different tab bar
depending on which scope it is within.

To facilitate this, a new `render_tab_bar` helper has been added. It
contains the logic to determine which bar should be rendered for the
scope.