Add a list of all cluster components in the navbar
This is the first step at bringing all the components into the nav-bar.
The tricky bit for components is that they are grouped.

For consistency, the components nav button displays the entire list. In
order for this to work, the components table partial needed to be
updated. It now accepts a list from the controller (otherwise it
defaults to the previous behaviour).

This is a step towards the controller determining what is displayed
instead of the view.