Ensure MW gracefully invalid after bad transition
Previously, we would blindly carry on trying to add the transition
comment after an invalid transition (e.g. a transition which requires a
user when one was not given). This would normally then fail later in a
non-descriptive way due to invariants we expect in the transition model
not being true (which is at least better than the alternative, of adding
the transition comment even though the transition was not successful).

A maintenance window is already invalid if it has an invalid transition,
so we just need to check this before attempting to add the comment.