Use the type param to select componets to index
The purpose of the drop down menu is to select which
`component_groups_types` should be displayed.

This is done with a `type` parameter that is sent with the request. This
is then converted into the corresponding `groups`. If no type is sent,
then all the groups are used.

This required reverting the `components_table` partial back to rendering groups.

The maintenance spec needed to be updated as the components are now on a
different page