Use consistent terminology for un/finished MWs
Previously we would sometimes refer to a MaintenanceWindow being
open/closed, and other times refer to it being finished/unfinished. This
commit changes this to consistently refer to this as
finished/unfinished, at least as far as I can see, in terms of both
method names and comments/test descriptions.

Finished/unfinished was chosen to refer to this concept as it seems less
ambiguous; a MaintenanceWindow is 'finished' once it has reached a state
which it cannot transition out of.