Use the tab bar as the header
This commit switches over to using the tab bar as the header for most
pages.

The only page that hasn't been switched is the cases. This is because of
how the cases use a partial that can not be readily adapted. This will
have to be done at a future date.

In order for the rest of the website to work, the content still needs to
be rendered even if it isn't in a cluster scope (and thus does not have
a tab bar).

To allow this, the content is rendered with a new `card` partial. It is
simply a card div that yields to accept the content.