Fix invalid Cases if change model support type
We validate the relationship between a Case's Issue and it's associated
model (Component, Service, or Cluster), as some Issues should only be
allowed to have Cases created associated with managed or advice models.
This works fine when the Case is first created, but leads to an issue
where the Case becomes invalid at a later date if the associated model's
support type is changed so that the same Case would not be allowed if it
was created today.

This happened to surface when trying to view the Case archive page for a
Site as an admin, as behind the scenes this checks the current RT status
for Cases which have not been completed; this was then causing the Case
to be updated leading this validation error to then be noticed for Cases
which had become invalid since their creation.

The actual fix is quite simple although the diff is large: just split
the Case validations into two files, depending on whether they are
invariants which should always hold or if they should just be checks
which should be run on Case creation.

The large diff results from splitting our custom Validator loosely in
half (which could still be improved as this has some duplication),
documenting why this was done, and adding test coverage for the
different potential issues this split resolves.