Move the bulk of table partial into cases index
Previously the `cases_table` was used to display the list of cases. This
will eventually be completely replaced with the index template. As the
table is still currently used in other places, the partial now renders
the index template.

The case controller also sets whether it show display in archive mode or
not.