Use the index case action for both regular and archive
The index action can now be used to show both the regular list of cases
and all the archived cases. This is done by sending a special
`archive=true` parameter with the request.

The default is not to display the archived cases. To accommodate this,
the spec had to be updated.

The title to the page has also been changed to be just `Manage Support
Cases` as the distinction was arbitrary.