Test the creation of a dropdown nav item
This confirms that a dropdown can be created correctly. The admin
dropdown has not been tested as it is intended for removal.

The `current_user` method has been mocked directly to be the user.

The `rendered` method has been wrapped to return a `Capybara` element.
This allows the finder methods to be used with it.

A new `icon_span` helper has been created to facilitate the creation of
links with icons. It is intended that dropdown items will also support
icons.