Track requested start/end in MW transition records
The values for `requested_start` and `requested_end` may soon
potentially change for a MaintenanceWindow at different points; in
particular a site contact will soon be able to  change these values when
the confirm some maintenance. Therefore we now want to track these
values as they change in the MaintenanceWindow transition model so that
we do not lose this potentially useful information.